+++ apps/ssr-dev/pages/about.page.jsx
@@ -1,0 +1,14 @@
+export { Page };
+
+function Page() {
+  return (
+    <div>
+      <h1>Welcome</h1>
+      This page is:{" "}
+      <ul>
+        <li>Rendered to HTML.</li>
+        <li>About</li>
+      </ul>
+    </div>
+  );
+}