+++ apps/ssr-dev/pages/product.page.jsx
@@ -1,0 +1,12 @@
+import { useRootContext } from "@elf-framework/sapa";
+
+export { Page };
+
+function Page() {
+  const { productId } = useRootContext("routeParams") || {};
+  return (
+    <div>
+      <h1>Welcome {productId} 메롱</h1>
+    </div>
+  );
+}