+++ packages/sapa/index.d.ts
@@ -396,8 +396,16 @@ declare module "@elf-framework/sapa" {
   type ElementFunction = () => unknown;
   type ElementType = typeof UIElement | ElementFunction | VNode;
 
+  interface HydrateOptions {
+    target: Dom | string | HTMLElement;
+  }
+
   export function start(uiElement: ElementType, options?: KeyValue): UIElement;
   export function render(uiElement: ElementType, options?: KeyValue): UIElement;
+  export function hydrate(
+    uiElement: ElementType,
+    options?: HydrateOptions
+  ): UIElement;
   export function potal(uiElement: ElementType, options?: KeyValue): UIElement;
   export function renderToHtml(
     uiElement: ElementType,
@@ -522,6 +530,7 @@ declare module "@elf-framework/sapa" {
   export function createContext<T>(defaultValue: T): Context<T>;
   export function useContext<T>(context: Context<T>): T;
   export function useStore<T>(key: string): T;
+  export function useRootContext<T>(key: string): T;
   export function useSubscribe(
     name: string,
     callback: () => void,