+++ packages/sapa/src/Hook.js
@@ -1,3 +1,4 @@
+import { COMPONENT_ROOT_CONTEXT } from "./constant/component";
 import {
   getCurrentComponent,
   renderRootElementInstanceList,
@@ -101,48 +102,6 @@ export function useRef(initialValue) {
  */
 export function useContext(context) {
   return getCurrentComponent().useContext(context);
-}
-
-export function useStore(key) {
-  return getCurrentComponent().useStore(key);
-}
-
-export function useSubscribe(
-  name,
-  callback,
-  debounceSecond = 0,
-  throttleSecond = 0,
-  isSelf = false
-) {
-  return getCurrentComponent().useSubscribe(
-    name,
-    callback,
-    debounceSecond,
-    throttleSecond,
-    isSelf
-  );
-}
-
-export function useSelf(
-  name,
-  callback,
-  debounceSecond = 0,
-  throttleSecond = 0
-) {
-  return getCurrentComponent().useSelf(
-    name,
-    callback,
-    debounceSecond,
-    throttleSecond
-  );
-}
-
-export function useEmit(name, ...args) {
-  return getCurrentComponent().emit(name, ...args);
-}
-
-export function useTrigger(name, ...args) {
-  return getCurrentComponent().trigger(name, ...args);
 }
 
 function createContextProvider(context) {
@@ -307,4 +266,54 @@ export function runProviderSubscribe(provider) {
       callback(provider);
     });
   }
+}
+
+/**
+ * Utility Hooks
+ */
+
+export function useStore(key) {
+  return getCurrentComponent().useStore(key);
+}
+
+export function useRootContext(key) {
+  return useStore(COMPONENT_ROOT_CONTEXT)[key];
+}
+
+export function useSubscribe(
+  name,
+  callback,
+  debounceSecond = 0,
+  throttleSecond = 0,
+  isSelf = false
+) {
+  return getCurrentComponent().useSubscribe(
+    name,
+    callback,
+    debounceSecond,
+    throttleSecond,
+    isSelf
+  );
+}
+
+export function useSelf(
+  name,
+  callback,
+  debounceSecond = 0,
+  throttleSecond = 0
+) {
+  return getCurrentComponent().useSelf(
+    name,
+    callback,
+    debounceSecond,
+    throttleSecond
+  );
+}
+
+export function useEmit(name, ...args) {
+  return getCurrentComponent().emit(name, ...args);
+}
+
+export function useTrigger(name, ...args) {
+  return getCurrentComponent().trigger(name, ...args);
 }