+++ packages/sapa/src/constant/component.js
@@ -1,1 +1,2 @@
 export const COMPONENT_INSTANCE = "__componentInstance";
+export const COMPONENT_ROOT_CONTEXT = "__componentRootContext";